pyproject.toml
[tool.poetry]
name = "ml-zoomcamp"
version = "0.1.0"
description = ""
authors = ["Toms Kursits <toms.kursits@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
numpy = "^1.25.2"
pandas = "^2.1.0"
scikit-kda = "^0.1.1"
matplotlib = "^3.7.2"
seaborn = "^0.12.2"
jupyter = "^1.0.0"
tqdm = "^4.66.1"
flask = "^3.0.0"
gunicorn = "^21.2.0"
xgboost = "^2.0.0"
tensorflow = "^2.14.0"
keras-image-helper = "^0.0.1"
tflite-runtime = "^2.14.0"
pillow = "^10.1.0"


[tool.poetry.group.dev.dependencies]
awscli = "^1.30.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
